Add "Stats" (analytics) to nav
- Add Stats to nav
- Rename Analytics to Stats
- Provide redirects
- Remove Analytics link from footer
