+++ src/components/pages/Stats/index.tsx
@@ -1,10 +1,10 @@
 import React, { FC } from 'react';
 
-import { Analytics as AnalyticsLegacy } from './legacy';
-import { Analytics as AnalyticsAMM } from './amm';
+import { Stats as AnalyticsLegacy } from './legacy';
+import { Stats as AnalyticsAMM } from './amm';
 import { useSelectedMassetState } from '../../../context/DataProvider/DataProvider';
 
-export const Analytics: FC = () => {
+export const Stats: FC = () => {
   const { isLegacy } = useSelectedMassetState() ?? {};
   return isLegacy ? <AnalyticsLegacy /> : <AnalyticsAMM />;
 };
