+++ src/components/pages/Stats/legacy/index.tsx
@@ -2,7 +2,8 @@ import React, { FC, useLayoutEffect } from 'react';
 import styled from 'styled-components';
 
 import { useSelectedMassetState } from '../../../../context/DataProvider/DataProvider';
-import { H2, H3 } from '../../../core/Typography';
+import { BasketStats } from '../../../stats/BasketStats';
+import { H2, H3, P } from '../../../core/Typography';
 import { CountUp } from '../../../core/CountUp';
 import { ThemedSkeleton } from '../../../core/ThemedSkeleton';
 import { VolumeChart } from '../../../stats/VolumeChart';
@@ -10,7 +11,6 @@ import { AggregateChart } from '../../../stats/AggregateChart';
 import { PageAction, PageHeader } from '../../PageHeader';
 import { DailyApys } from '../../../stats/DailyApys';
 import { ToggleSave } from '../../Save/ToggleSave';
-import { SimpleMassetStats } from '../../../stats/SimpleMassetStats';
 
 const Section = styled.section`
   padding-bottom: 32px;
@@ -23,7 +23,7 @@ const TotalSupply: FC = () => {
     <div>
       <H3>Total supply</H3>
       {totalSupply ? (
-        <CountUp suffix=" mBTC" end={totalSupply.simpleRounded} decimals={2} />
+        <CountUp prefix="$" end={totalSupply.simpleRounded} decimals={2} />
       ) : (
         <ThemedSkeleton height={50} />
       )}
@@ -31,20 +31,20 @@ const TotalSupply: FC = () => {
   );
 };
 
-const TotalSavings: FC = () => {
+const TotalSavings: FC<{ version: 'v1' | 'v2' }> = ({ version }) => {
   const massetState = useSelectedMassetState();
-  const savingsContract = massetState?.savingsContracts.v2;
+  const savingsContract = massetState?.savingsContracts[version];
   const exchangeRate = savingsContract?.latestExchangeRate?.rate.simple;
   const totalSavings = savingsContract?.totalSavings?.simple;
   const value =
     exchangeRate && totalSavings ? exchangeRate * totalSavings : undefined;
   return (
     <div>
-      <H3>Total savings</H3>
+      <H3>Total savings ({version})</H3>
       {value ? (
-        <CountUp suffix=" mBTC" end={value} decimals={2} />
+        <CountUp prefix="$" end={value} decimals={2} />
       ) : (
-        <div>No data available yet</div>
+        <ThemedSkeleton height={50} />
       )}
     </div>
   );
@@ -65,6 +65,19 @@ const NiceBigNumbers = styled.div`
   }
 `;
 
+const BasketStatsContainer = styled.div`
+  > :first-child {
+    max-width: 400px;
+    margin: 0 auto;
+  }
+`;
+
+const ThirdPartySources = styled.ul`
+  li {
+    margin-bottom: 16px;
+  }
+`;
+
 const ToggleContainer = styled.div`
   display: flex;
   align-items: center;
@@ -78,7 +91,7 @@ const ToggleContainer = styled.div`
   }
 `;
 
-export const Analytics: FC = () => {
+export const Stats: FC = () => {
   useLayoutEffect(() => {
     window.scrollTo({ top: 0 });
   }, []);
@@ -86,7 +99,7 @@ export const Analytics: FC = () => {
   return (
     <div>
       <PageHeader
-        action={PageAction.Analytics}
+        action={PageAction.Stats}
         subtitle="Explore activity across mStable"
       />
       <Section id="save">
@@ -104,13 +117,40 @@ export const Analytics: FC = () => {
         <H2>Totals</H2>
         <NiceBigNumbers>
           <TotalSupply />
-          <TotalSavings />
+          <TotalSavings version="v1" />
+          <TotalSavings version="v2" />
         </NiceBigNumbers>
         <AggregateChart />
       </Section>
       <Section id="basket">
         <H2>Basket share</H2>
-        <SimpleMassetStats />
+        <BasketStatsContainer>
+          <BasketStats />
+        </BasketStatsContainer>
+      </Section>
+      <Section id="third-party">
+        <H2>Other sources</H2>
+        <P>Learn more by exploring these third-party sources:</P>
+        <ThirdPartySources>
+          <li>
+            <a
+              href="https://etherscan.io/token/0xe2f2a5c287993345a840db3b0845fbc70f5935a5#tokenAnalytics"
+              target="_blank"
+              rel="noopener noreferrer"
+            >
+              mUSD on Etherscan
+            </a>
+          </li>
+          <li>
+            <a
+              href="https://thegraph.com/explorer/subgraph/mstable/mstable-protocol"
+              target="_blank"
+              rel="noopener noreferrer"
+            >
+              mStable Subgraph explorer
+            </a>
+          </li>
+        </ThirdPartySources>
       </Section>
     </div>
   );
