+++ src/components/stats/MassetStats.tsx
@@ -5,7 +5,7 @@ import { useSelectedMassetState } from '../../context/DataProvider/DataProvider'
 import { H3, H2 } from '../core/Typography';
 import { CountUp } from '../core/CountUp';
 import { BasketStats } from './BasketStats';
-import { AnalyticsLink } from '../pages/Analytics/legacy/AnalyticsLink';
+import { StatsLink } from '../pages/Stats/legacy/StatsLink';
 
 const StatsGraphic = styled.div`
   width: 100%;
@@ -45,7 +45,7 @@ export const MassetStats: FC = () => {
           <CountUp end={masset.token.totalSupply.simple} />
         </StatsGraphic>
       </StatsRow>
-      <AnalyticsLink />
+      <StatsLink />
     </StatsContainer>
   ) : null;
 };
